+++ include/barry/barray-iterator.hpp
@@ -10,14 +10,14 @@ template<typename Cell_Type, typename Data_Type>
 class ConstBArrayRowIter {
 public:
     
-    uint current_row, current_col;
+    size_t current_row, current_col;
     typename Row_type<Cell_Type>::const_iterator iter;
     const BArray<Cell_Type,Data_Type> * Array;
     
     ConstBArrayRowIter(const BArray<Cell_Type,Data_Type> * Array_) : Array(Array_) {
         
         // Finding the first entry of the iterator
-        for (uint i = 0u; i < Array->nrow(); ++i)
+        for (size_t i = 0u; i < Array->nrow(); ++i)
             if (A_ROW(i).size() != 0u) {
                 iter = A_ROW(i).begin();
                 break;
